RGB24 & friends: deprecate construction with & comparison to UInt32
The problem is that Gray24(x) should interpret x according to its value (between 0 and 1), and having a special case for UInt32 is confusing. The other 24/32 constructors have fallbacks for gray->RGB conversion and supplying alpha=1, so they face the same problem.

Deprecate in favor of explicitly using reinterpret, which is very clear.

Among other things, this fixes several StackOverflows.