RFC: `eltype`->`eltypec`, `length`->`lengthc`
Long ago this package decided to treat colorants as "scalars"
and not support Julia's standard iteration interface over color
channels. We implemented `mapc`, `reducec`, and
`mapreducec` to fill the need to sometimes write code
that iterates over channels.

Otherwise Colorants have mostly been thought of similar
to how one would typically work with numbers, albeit
with some caution to address subtleties of color math.

However, there are violations of the standard number iterface: for numbers, typically we expect
`eltype(T) == T` and `length(x) == 1`, neither of which
were true for some Colorants.  Likewise, the core
iterator trait `Base.IteratorSize` returns `HasLength()`
for `Colorant`s but `HasShape{0}()` for numbers.
Finally, `iterate(c)` currently throws an error, but
`iterate(2)` does not. (Numbers are iterable in Julia to support
generic code for which you can pass either `dims = (2,3,5)`,
`dims = [2,3,5]`, or `dims = 2`.)

This PR takes a daring step towards treating Colorants just like
numbers with respect to the entire iterator interface.
It deprecates `eltype(C)` in favor of `eltypec(C)`,
and `length(c)` in favor of `lengthc(c)`.
To complete the transition, after a suitable period we should
return`eltype(C) === C`, `length(c) == 1`, and support iteration
of the same style as for Numbers.

Needless to say, this is a scary change with the potential for a
lot of downstream code churn.