Add `zero` & `oneunit` for more types
This moves the `zero` and `oneunit` for  some of grays and rgbs
that were defined in ColorVectorSpace to ColorTypes.
For `zero`, a color with all components zero is considered reasonable
as an additive identity in most color spaces.
On the other hand, it is difficult to define `oneunit` in general.
Therefore, oneunit is defined only for specific color spaces.

This does not change the definition of `one`.

Co-authored-by: Tim Holy <tim.holy@gmail.com>